Fix `ring_theory/norm.lean` (timeout fix by helping the elaborator)
This seems to be a consequence of conflicting `field` instances for the subtype associated to an intermediate field. This is fixed in the follow-up PR by removing the `intermediate_field.to_field` instance, which is postponed in this PR since that requires a bit more cleanup of the dependencies.

An alternative would be to have a `local attribute [-instance] subfield_class.to_field`, which would need to be removed in a follow-up PR. I've opted for this change since it doesn't need any follow-ups.